cc: scale a pointer by the object's byte size, not by a saturated bit count
`&a[1][0] - &a[0][0]` on a `char[3][600000000]` answered 536870911. So did
the stride of an array of a 600 MB struct, and `p + 1` on a pointer to one.
`sizeof` was right throughout, so the sizes agreed with gcc while the
addresses did not -- a wrong answer with no diagnostic, at every optimization
level, on both targets.

536870911 is `u32::MAX / 8`, the value `size_bits` saturates to. It answers a
*value* width -- what an instruction operand holds, bounded at 128 bits -- and
an aggregate past `u32::MAX` bits has none. While the object-size bound was
that same number the saturation was unreachable, because the parser refused
any type that could reach it, so deriving a byte count as `size_bits / 8` was
safe by accident. Raising the bound to `u64::MAX / 8` made it reachable and
left thirty-five sites deriving byte counts from a number that no longer fits
one.

Every one of them asks `size_bytes` now: array indexing and pointer
arithmetic, stacked-argument and frame-slot sizes on both targets, static
initializer layout, the atomic access width, and a complex value's imaginary
half. Two were spelled `div_ceil(8)` rather than `/ 8` and were missed by the
first pass, which is the argument for making the unit a type rather than a
convention -- recorded in TODO.md with what that would take, and why widening
`size_bits` to `u64` is not it.

`bitfield_storage_type` takes a byte count, so it takes the type an object
size is counted in.

Verified against gcc on x86-64 natively and on aarch64 under qemu. The
regression test asserts on the emitted scale factor rather than running: the
element size is computed in `ir/linearize.rs` before any backend sees it, and
no test should ask its machine for gigabytes to observe a constant. It fails
on the previous commit.

+++ cc/ir/linearize_emit.rs
@@ -376,7 +376,7 @@ impl<'a> super::linearize::Linearizer<'a> {
         }
 
         // Determine storage type based on storage unit size
-        let storage_type = self.bitfield_storage_type(storage_size);
+        let storage_type = self.bitfield_storage_type(storage_size as usize);
         let storage_bits = storage_size * 8;
 
         // 1. Load the entire storage unit
@@ -691,7 +691,7 @@ impl<'a> super::linearize::Linearizer<'a> {
         }
 
         // Determine storage type based on storage unit size
-        let storage_type = self.bitfield_storage_type(storage_size);
+        let storage_type = self.bitfield_storage_type(storage_size as usize);
         let storage_bits = storage_size * 8;
 
         // 1. Load current storage unit value
